breaking out html input stories
