+++ html/base/inputs/Checkbox.stories.js
@@ -1,0 +1,90 @@
+import { useEffect } from '@storybook/addons';
+import { requiredTick } from '../requiredTick';
+
+export default {
+  title: 'Components/Input/Checkbox',
+  decorators: [
+    story => `<div class="sprk-o-Box">${story()}</div>`,
+  ],
+  parameters: {
+    info: `
+##### For design and usage information check out the [documentation.](https://spark-docs.netlify.com/using-spark/components/input)
+    `,
+  },
+};
+
+export const checkbox = () => {
+  useEffect(() => {
+    requiredTick();
+  }, []);
+
+  return `
+    <div class="sprk-b-InputContainer">
+      <fieldset class="sprk-b-Fieldset">
+        <legend class="sprk-b-Legend">
+          <label class="sprk-b-Label">
+            Checkbox Group Label
+          </label>
+        </legend>
+
+        <div class="sprk-b-SelectionContainer">
+          <input
+            id="checkbox-1"
+            data-id="checkbox-1"
+            type="checkbox"
+            aria-describedby="checkbox--error-container"
+          >
+
+          <label
+            for="checkbox-1"
+            class="sprk-b-Label sprk-b-Label--inline"
+          >
+            Checkbox Item 1
+          </label>
+        </div>
+
+        <div class="sprk-b-SelectionContainer">
+          <input
+            id="checkbox-2"
+            data-id="checkbox-2"
+            type="checkbox"
+            aria-describedby="checkbox--error-container"
+          >
+
+          <label
+            for="checkbox-2"
+            class="sprk-b-Label sprk-b-Label--inline"
+          >
+            Checkbox Item 2
+          </label>
+        </div>
+
+        <div class="sprk-b-SelectionContainer">
+          <input
+            id="checkbox-3"
+            data-id="checkbox-3"
+            type="checkbox"
+            aria-describedby="checkbox--error-container"
+          >
+
+          <label
+            for="checkbox-3"
+            class="sprk-b-Label sprk-b-Label--inline"
+          >
+            Checkbox Item 3
+          </label>
+        </div>
+      </fieldset>
+
+      <div
+        class="sprk-b-ErrorContainer"
+        id="checkbox--error-container"
+      >
+      </div>
+    </div>
+  `;
+};
+
+checkbox.story = {
+  name: 'Default',
+};
